% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\docType{data}
\name{eppa_colors}
\alias{eppa_colors}
\title{Color palette for MIT's EPPA model}
\format{An object of class \code{character} of length 18.}
\usage{
eppa_colors
}
\description{
This palette should be used for plots by region (whether maps, line plots, or
other types) to ensure consistency across plots and publications.
}
\keyword{datasets}
