% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadDefaultProjectSettings}
\alias{loadDefaultProjectSettings}
\title{Load the default project file into the settings}
\usage{
loadDefaultProjectSettings()
}
\description{
Returns the settings from the default project file
}
