% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadDefaultRegionColors}
\alias{loadDefaultRegionColors}
\title{Load the default region colors}
\usage{
loadDefaultRegionColors()
}
\description{
Returns the region colors from the default project file
}
