% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadDefaultSectorColors}
\alias{loadDefaultSectorColors}
\title{Load the default sector colors}
\usage{
loadDefaultSectorColors()
}
\description{
Returns the sector colors from the default project file
}
