% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadProjectSettings}
\alias{loadProjectSettings}
\title{Load a file into the settings}
\usage{
loadProjectSettings(file)
}
\arguments{
\item{proj}{Path to the project file}
}
\description{
Returns the settings from the project file
}
