% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{plotTime}
\alias{plotTime}
\title{Plot values over time as a bar chart}
\usage{
plotTime(
  prjdata,
  plot_type,
  query,
  scen,
  diffscen,
  subcatvar,
  filter,
  rgns,
  regionColors,
  sectorColors
)
}
\arguments{
\item{prjdata}{A project data structure}

\item{plot_type}{The type to plot: stacked or line}

\item{query}{Name of the query to plot}

\item{scen}{Name of the scenario to plot}

\item{diffscen}{Name of the difference scenario, or NULL if none}

\item{subcatvar}{Variable to use for subcategories in the plot}

\item{filter}{If TRUE, then filter to regions in the rgns argument}

\item{rgns}{Regions to filter to, if filter is TRUE.}

\item{regionColors}{Region colors to use, if plotting by region}

\item{sectorColors}{Sector colors to use, if plotting by sector}
}
\description{
Plot values over time as a bar chart
}
