% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-helpers.R
\name{loadProject2}
\alias{loadProject2}
\title{Load a file into the UI}
\usage{
loadProject2(proj)
}
\arguments{
\item{proj}{Path to the project file}
}
\description{
Returns the data from the project file, if valid
}
